Some updates to the interpolators and benchmarkers.

+++ code/essentials/sources/vitypesizeconverter.cpp
@@ -3,7 +3,7 @@
 template <typename T1, typename T2>
 int ViTypeSizeConverter<T1, T2>::convert(int size)
 {
-	return size * (sizeof(T2) / float(sizeof(T1)));
+	return size * (sizeof(T2) / double(sizeof(T1)));
 }
 
 #endif
